// clients/js-kit/src/generated/types/baseRoyalties.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getArrayDecoder,
  getArrayEncoder,
  getStructDecoder,
  getStructEncoder,
  getU16Decoder,
  getU16Encoder,
  type Codec,
  type Decoder,
  type Encoder,
} from '@solana/kit';
import {
  getBaseRuleSetDecoder,
  getBaseRuleSetEncoder,
  getCreatorDecoder,
  getCreatorEncoder,
  type BaseRuleSet,
  type BaseRuleSetArgs,
  type Creator,
  type CreatorArgs,
} from '.';

export type BaseRoyalties = {
  basisPoints: number;
  creators: Array<Creator>;
  ruleSet: BaseRuleSet;
};

export type BaseRoyaltiesArgs = {
  basisPoints: number;
  creators: Array<CreatorArgs>;
  ruleSet: BaseRuleSetArgs;
};

export function getBaseRoyaltiesEncoder(): Encoder<BaseRoyaltiesArgs> {
  return getStructEncoder([
    ['basisPoints', getU16Encoder()],
    ['creators', getArrayEncoder(getCreatorEncoder())],
    ['ruleSet', getBaseRuleSetEncoder()],
  ]);
}

export function getBaseRoyaltiesDecoder(): Decoder<BaseRoyalties> {
  return getStructDecoder([
    ['basisPoints', getU16Decoder()],
    ['creators', getArrayDecoder(getCreatorDecoder())],
    ['ruleSet', getBaseRuleSetDecoder()],
  ]);
}

export function getBaseRoyaltiesCodec(): Codec<
  BaseRoyaltiesArgs,
  BaseRoyalties
> {
  return combineCodec(getBaseRoyaltiesEncoder(), getBaseRoyaltiesDecoder());
}